// src/app/models/Achat.ts
export class Achat {
    idAAchat: number;
    deviseAED: string;
    deviseEUR: string;
    deviseLYD: string;
    deviseBHD: string;
    deviseSAR: string;
    deviseCAD: string;
    deviseDKK: string;
    deviseUSD: string;
    deviseGBP: string;
    deviseJPY: string;
    deviseNOK: string;
    deviseSEK: string;
    deviseCHF: string;
    deviseKWD: string;
    deviseQAR: string;
    deviseCNY: string;
    nomBanque: string;
    dateDevise: Date;
  }